+++ primefaces/src/main/java/org/primefaces/model/DialogFrameworkOptions.java
@@ -29,7 +29,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class DialogFrameworkOptions {
-
     private String widgetVar;
     private boolean modal = false;
     private boolean resizable = true;
@@ -38,6 +37,27 @@ public class DialogFrameworkOptions {
     private String height;
     private String contentWidth;
     private String contentHeight;
+    private boolean closable = true;
+    private boolean includeViewParams = false;
+    private String headerElement;
+    private boolean minimizable = false;
+    private boolean maximizable = false;
+    private boolean closeOnEscape = false;
+    private Integer minWidth;
+    private Integer minHeight;
+    private String appendTo;
+    private boolean dynamic = false;
+    private String showEffect;
+    private String hideEffect;
+    private String position;
+    private boolean fitViewport = false;
+    private boolean responsive = false;
+    private String focus;
+    private String onShow;
+    private String onHide;
+    private boolean blockScroll = false;
+    private String styleClass;
+    private String iframeStyleClass;
 
     public String getWidgetVar() {
         return widgetVar;
@@ -103,6 +123,174 @@ public class DialogFrameworkOptions {
         this.contentHeight = contentHeight;
     }
 
+    public boolean isClosable() {
+        return closable;
+    }
+
+    public void setClosable(boolean closable) {
+        this.closable = closable;
+    }
+
+    public boolean isIncludeViewParams() {
+        return includeViewParams;
+    }
+
+    public void setIncludeViewParams(boolean includeViewParams) {
+        this.includeViewParams = includeViewParams;
+    }
+
+    public String getHeaderElement() {
+        return headerElement;
+    }
+
+    public void setHeaderElement(String headerElement) {
+        this.headerElement = headerElement;
+    }
+
+    public boolean isMinimizable() {
+        return minimizable;
+    }
+
+    public void setMinimizable(boolean minimizable) {
+        this.minimizable = minimizable;
+    }
+
+    public boolean isMaximizable() {
+        return maximizable;
+    }
+
+    public void setMaximizable(boolean maximizable) {
+        this.maximizable = maximizable;
+    }
+
+    public boolean isCloseOnEscape() {
+        return closeOnEscape;
+    }
+
+    public void setCloseOnEscape(boolean closeOnEscape) {
+        this.closeOnEscape = closeOnEscape;
+    }
+
+    public Integer getMinWidth() {
+        return minWidth;
+    }
+
+    public void setMinWidth(Integer minWidth) {
+        this.minWidth = minWidth;
+    }
+
+    public Integer getMinHeight() {
+        return minHeight;
+    }
+
+    public void setMinHeight(Integer minHeight) {
+        this.minHeight = minHeight;
+    }
+
+    public String getAppendTo() {
+        return appendTo;
+    }
+
+    public void setAppendTo(String appendTo) {
+        this.appendTo = appendTo;
+    }
+
+    public boolean isDynamic() {
+        return dynamic;
+    }
+
+    public void setDynamic(boolean dynamic) {
+        this.dynamic = dynamic;
+    }
+
+    public String getShowEffect() {
+        return showEffect;
+    }
+
+    public void setShowEffect(String showEffect) {
+        this.showEffect = showEffect;
+    }
+
+    public String getHideEffect() {
+        return hideEffect;
+    }
+
+    public void setHideEffect(String hideEffect) {
+        this.hideEffect = hideEffect;
+    }
+
+    public String getPosition() {
+        return position;
+    }
+
+    public void setPosition(String position) {
+        this.position = position;
+    }
+
+    public boolean isFitViewport() {
+        return fitViewport;
+    }
+
+    public void setFitViewport(boolean fitViewport) {
+        this.fitViewport = fitViewport;
+    }
+
+    public boolean isResponsive() {
+        return responsive;
+    }
+
+    public void setResponsive(boolean responsive) {
+        this.responsive = responsive;
+    }
+
+    public String getFocus() {
+        return focus;
+    }
+
+    public void setFocus(String focus) {
+        this.focus = focus;
+    }
+
+    public String getOnShow() {
+        return onShow;
+    }
+
+    public void setOnShow(String onShow) {
+        this.onShow = onShow;
+    }
+
+    public String getOnHide() {
+        return onHide;
+    }
+
+    public void setOnHide(String onHide) {
+        this.onHide = onHide;
+    }
+
+    public boolean isBlockScroll() {
+        return blockScroll;
+    }
+
+    public void setBlockScroll(boolean blockScroll) {
+        this.blockScroll = blockScroll;
+    }
+
+    public String getStyleClass() {
+        return styleClass;
+    }
+
+    public void setStyleClass(String styleClass) {
+        this.styleClass = styleClass;
+    }
+
+    public String getIframeStyleClass() {
+        return iframeStyleClass;
+    }
+
+    public void setIframeStyleClass(String iframeStyleClass) {
+        this.iframeStyleClass = iframeStyleClass;
+    }
+
     // TODO: https://primefaces.github.io/primefaces/12_0_0/#/core/dialogframework
 
     public Map<String, Object> toMap() {
@@ -126,8 +314,51 @@ public class DialogFrameworkOptions {
         if (LangUtils.isNotBlank(contentHeight)) {
             opts.put("contentHeight", contentHeight);
         }
-
-        // TODO: https://primefaces.github.io/primefaces/12_0_0/#/core/dialogframework
+        opts.put("closable", closable);
+        opts.put("includeViewParams", includeViewParams);
+        if (LangUtils.isNotBlank(headerElement)) {
+            opts.put("headerElement", headerElement);
+        }
+        opts.put("minimizable", minimizable);
+        opts.put("maximizable", maximizable);
+        opts.put("closeOnEscape", closeOnEscape);
+        if (minWidth != null) {
+            opts.put("minWidth", minWidth);
+        }
+        if (minHeight != null) {
+            opts.put("minHeight", minHeight);
+        }
+        if (LangUtils.isNotBlank(appendTo)) {
+            opts.put("appendTo", appendTo);
+        }
+        opts.put("dynamic", dynamic);
+        if (LangUtils.isNotBlank(showEffect)) {
+            opts.put("showEffect", showEffect);
+        }
+        if (LangUtils.isNotBlank(hideEffect)) {
+            opts.put("hideEffect", hideEffect);
+        }
+        if (LangUtils.isNotBlank(position)) {
+            opts.put("position", position);
+        }
+        opts.put("fitViewport", fitViewport);
+        opts.put("responsive", responsive);
+        if (LangUtils.isNotBlank(focus)) {
+            opts.put("focus", focus);
+        }
+        if (LangUtils.isNotBlank(onShow)) {
+            opts.put("onShow", onShow);
+        }
+        if (LangUtils.isNotBlank(onHide)) {
+            opts.put("onHide", onHide);
+        }
+        opts.put("blockScroll", blockScroll);
+        if (LangUtils.isNotBlank(styleClass)) {
+            opts.put("styleClass", styleClass);
+        }
+        if (LangUtils.isNotBlank(iframeStyleClass)) {
+            opts.put("iframeStyleClass", iframeStyleClass);
+        }
 
         return opts;
     }
@@ -224,7 +455,211 @@ public class DialogFrameworkOptions {
             return this;
         }
 
-        // TODO: https://primefaces.github.io/primefaces/12_0_0/#/core/dialogframework
+        /**
+         * Whether the dialog can be closed or not.
+         * @param closable
+         * @return
+         */
+        public DialogFrameworkOptions.Builder closable(boolean closable) {
+            options.setClosable(closable);
+            return this;
+        }
+
+        /**
+         * When enabled, includes the view parameters.
+         * @param includeViewParams
+         * @return
+         */
+        public DialogFrameworkOptions.Builder includeViewParams(boolean includeViewParams) {
+            options.setIncludeViewParams(includeViewParams);
+            return this;
+        }
+
+        /**
+         * Client id of the element to display inside header.
+         * @param headerElement
+         * @return
+         */
+        public DialogFrameworkOptions.Builder headerElement(String headerElement) {
+            options.setHeaderElement(headerElement);
+            return this;
+        }
+
+        /**
+         * Makes dialog minimizable.
+         * @param minimizable
+         * @return
+         */
+        public DialogFrameworkOptions.Builder minimizable(boolean minimizable) {
+            options.setMinimizable(minimizable);
+            return this;
+        }
+
+        public DialogFrameworkOptions.Builder maximizable(boolean maximizable) {
+            options.setMaximizable(maximizable);
+            return this;
+        }
+
+        /**
+         * Whether the dialog can be closed with escape key.
+         * @param closeOnEscape
+         * @return
+         */
+        public DialogFrameworkOptions.Builder closeOnEscape(boolean closeOnEscape) {
+            options.setCloseOnEscape(closeOnEscape);
+            return this;
+        }
+
+        /**
+         * Minimum width of a resizable dialog.
+         * @param minWidth
+         * @return
+         */
+        public DialogFrameworkOptions.Builder minWidth(Integer minWidth) {
+            options.setMinWidth(minWidth);
+            return this;
+        }
+
+        /**
+         * Minimum height of a resizable dialog.
+         * @param minHeight
+         * @return
+         */
+        public DialogFrameworkOptions.Builder minHeight(Integer minHeight) {
+            options.setMinHeight(minHeight);
+            return this;
+        }
+
+        /**
+         * Appends the dialog to the element defined by the given search expression.
+         * @param appendTo
+         * @return
+         */
+        public DialogFrameworkOptions.Builder appendTo(String appendTo) {
+            options.setAppendTo(appendTo);
+            return this;
+        }
+
+        /**
+         * Enables lazy loading of the content with ajax.
+         * @param dynamic
+         * @return
+         */
+        public DialogFrameworkOptions.Builder dynamic(boolean dynamic) {
+            options.setDynamic(dynamic);
+            return this;
+        }
+
+        /**
+         * Effect to use when showing the dialog
+         * @param showEffect
+         * @return
+         */
+        public DialogFrameworkOptions.Builder showEffect(String showEffect) {
+            options.setShowEffect(showEffect);
+            return this;
+        }
+
+        /**
+         * Effect to use when hiding the dialog
+         * @param hideEffect
+         * @return
+         */
+        public DialogFrameworkOptions.Builder hideEffect(String hideEffect) {
+            options.setHideEffect(hideEffect);
+            return this;
+        }
+
+        /**
+         * Defines where the dialog should be displayed
+         * @param position
+         * @return
+         */
+        public DialogFrameworkOptions.Builder position(String position) {
+            options.setPosition(position);
+            return this;
+        }
+
+        /**
+         * Dialog size might exceeed viewport if content is bigger than viewport in terms of height.
+         * fitViewport option automatically adjusts height to fit dialog within the viewport.
+         * @param fitViewport
+         * @return
+         */
+        public DialogFrameworkOptions.Builder fitViewport(boolean fitViewport) {
+            options.setFitViewport(fitViewport);
+            return this;
+        }
+
+        /**
+         * In responsive mode, dialog adjusts itself based on screen width.
+         * @param responsive
+         * @return
+         */
+        public DialogFrameworkOptions.Builder responsive(boolean responsive) {
+            options.setResponsive(responsive);
+            return  this;
+        }
+
+        /**
+         * Defines which component to apply focus by search expression.
+         * @param focus
+         * @return
+         */
+        public DialogFrameworkOptions.Builder focus(String focus) {
+            options.setFocus(focus);
+            return this;
+        }
+
+        /**
+         * Client side callback to execute when dialog is displayed.
+         * @param onShow
+         * @return
+         */
+        public DialogFrameworkOptions.Builder onShow(String onShow) {
+            options.setOnShow(onShow);
+            return this;
+        }
+
+        /**
+         * Client side callback to execute when dialog is hidden.
+         * @param onHide
+         * @return
+         */
+        public DialogFrameworkOptions.Builder onHide(String onHide) {
+            options.setOnHide(onHide);
+            return this;
+        }
+
+        /**
+         * Whether to block scrolling of the document when dialog is modal.
+         * @param blockScroll
+         * @return
+         */
+        public DialogFrameworkOptions.Builder blockScroll(boolean blockScroll) {
+            options.setBlockScroll(blockScroll);
+            return this;
+        }
+
+        /**
+         * One or more CSS classes for the dialog.
+         * @param styleClass
+         * @return
+         */
+        public DialogFrameworkOptions.Builder styleClass(String styleClass) {
+            options.setStyleClass(styleClass);
+            return this;
+        }
+
+        /**
+         * One or more CSS classes for the iframe within the dialog.
+         * @param iframeStyleClass
+         * @return
+         */
+        public DialogFrameworkOptions.Builder iframeStyleClass(String iframeStyleClass) {
+            options.setIframeStyleClass(iframeStyleClass);
+            return this;
+        }
 
         public DialogFrameworkOptions build() {
             return options;